[ fix ] Fix product card image animation
Corrected the animation of the product card image by applying the
`animate-onTheEarth` class to the container instead of the image
itself. This ensures that the animation is smooth and accurate.
Additionally, adjusted the gradient background to be placed
correctly using the `rotate-180` class.

Also, changed the condition for the `particular` class to
correctly check for the `_featured` property being equal to 1.