[ fix ] Improve Popup component UI and add close button
Improved the UI of the Popup component by making the content
width responsive and adding a close button.  The "do not show
again" checkbox was also improved to use translated text and
a second button was added allowing the user to navigate to the
relevant page instead of just closing the popup.  These changes
enhance the user experience and provide more control.