[add] Add unique ID to HtmlContainer
Added a unique ID to the HtmlContainer component.  This
prevents potential conflicts when multiple instances of the
component are rendered on the same page, particularly when
using JavaScript to interact with specific containers.  If an ID
is provided as a prop, it will be used; otherwise, a random ID
will be generated.